+++ apps/hostd/pages/_app.tsx
@@ -3,6 +3,7 @@ import { NextAppCsr } from '@siafoundation/design-system'
 import { AppProps } from 'next/app'
 import { Providers } from '../config/providers'
 import { routes } from '../config/routes'
+import { rootFontClasses } from '@siafoundation/fonts'
 
 export default function App({
   Component,
@@ -12,6 +13,7 @@ export default function App({
 }>) {
   return (
     <NextAppCsr
+      className={rootFontClasses}
       fallback={pageProps.fallback}
       passwordProtectRequestHooks
       lockRoutes={routes}
