+++ apps/website/pages/_document.tsx
@@ -1,14 +1,14 @@
 'use client'
 
-import { rootFontClasses } from '@siafoundation/fonts'
 import DefaultNextDocument, {
   Head,
   Html,
   Main,
   NextScript,
 } from 'next/document'
+import { rootFontClasses } from '@siafoundation/fonts'
 
-export class NextDocument extends DefaultNextDocument {
+export default class NextDocument extends DefaultNextDocument {
   override render() {
     return (
       <Html lang="en">
