+++ libs/fonts/package.json
@@ -5,7 +5,7 @@
   "license": "MIT",
   "dependencies": {
     "class-variance-authority": "^0.7.0",
-    "@next/font": "13.1.1"
+    "next": "13.5.2"
   },
   "types": "./src/index.d.ts"
 }
