+++ package.json
@@ -27,7 +27,6 @@
     "@ledgerhq/hw-transport-webhid": "^6.27.19",
     "@mdx-js/loader": "^2.1.1",
     "@mdx-js/react": "^2.1.1",
-    "@next/font": "13.1.1",
     "@next/mdx": "^12.1.6",
     "@notionhq/client": "^2.2.8",
     "@nx-tools/nx-container": "^5.0.3",
